ok im wanna to avoid node vtables

+++ src/drv/uart.h
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-void uart_begin(uint32_t baudrate);
+void uart_begin(uint16_t baudrate);
 void uart_write(uint8_t data);
 void uart_print_str(const char* str);
 bool uart_available();
